changed decimals to intgers in sequ models and updated frontend output

+++ server/db/models/orderProduct.js
@@ -12,7 +12,7 @@ const OrderProduct = db.define('orderProduct', {
     allowNull: false
   },
   price: {
-    type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.INTEGER,
     allowNull: false
   }
 })
